requirements.txt
pandas=2.0.3
matplotlib=3.7.4
scikit-learn=1.3.2
Flask==3.0.0
Flask-Limiter==3.5.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
cryptography==41.0.7
pytest==7.4.3
